requirements.txt
discord.py>=2.3.2
python-dotenv==1.0.1
SQLAlchemy==2.0.29
fastapi==0.111.0
uvicorn[standard]==0.29.0
alembic>=1.12.0
gspread>=5.12.0
google-auth-oauthlib>=1.2.0
